Refactor the accounts store module
