+++ src/store/accounts/actions.js
@@ -7,18 +7,11 @@ import EthWallet from 'ethereumjs-wallet';
 import { Wallet } from '@/class';
 import keystore from '@/utils/keystore';
 import {
-  SAVE_TOKENS,
-  SAVE_TRACKED_TOKENS,
-} from '@/store/tokens/mutations-types';
-import {
   SET_ADDRESS,
   SET_WALLET,
   ADD_WALLET,
   SET_HD_KEY,
   SET_BALANCE,
-  SET_SETTINGS,
-  SET_OTP_SETTINGS,
-  SET_EMAIL,
   ADD_ADDRESS,
 } from './mutations-types';
 
@@ -176,86 +169,14 @@ const updateBalance = async ({ commit, dispatch, state }) => {
   }
 };
 
-const updateSettings = async ({ commit, dispatch }, settings) => {
-  try {
-    commit(SET_SETTINGS, settings);
-    await userService.setSetting('settings', settings);
-  } catch (e) {
-    dispatch('errors/emitError', e, { root: true });
+const validatePassword = async ({ state, getters }, password) => {
+  let { wallet } = state;
+
+  if (getters.isPublicAccount) {
+    wallet = new Wallet(state.hdKey);
   }
-};
 
-const validatePassword = ({ state }, password) =>
-  state.wallet.validatePassword(password);
-
-const login = (ctx, email) => userService.login(email);
-
-const logout = async ({ commit, dispatch }) => {
-  try {
-    commit(SET_EMAIL, null);
-    await userService.logout();
-    window.location.reload();
-  } catch (e) {
-    dispatch('errors/emitError', e, { root: true });
-  }
-};
-
-const loginViaOTP = (ctx, { code, email }) =>
-  userService.loginViaOTP(code, email);
-
-const getOtpSettings = async ({ commit, dispatch }) => {
-  try {
-    const otpSettings = await userService.getOtpSettings();
-    commit(SET_OTP_SETTINGS, otpSettings);
-  } catch (e) {
-    dispatch('errors/emitError', e, { root: true });
-  }
-};
-
-const setOtpSettings = async ({ commit, dispatch }, { secret, code }) => {
-  try {
-    await userService.setOtpSettings(secret, code);
-    commit(SET_OTP_SETTINGS, { status: 'enabled' });
-  } catch (e) {
-    dispatch('errors/emitError', e, { root: true });
-  }
-};
-
-const deleteOtpSettings = async ({ commit, dispatch }, { code }) => {
-  try {
-    await userService.deleteOtpSettings(code);
-    commit(SET_OTP_SETTINGS, {});
-    await dispatch('getOtpSettings');
-  } catch (e) {
-    dispatch('errors/emitError', e, { root: true });
-  }
-};
-
-const setUserSettings = async ({ commit, dispatch }) => {
-  try {
-    const { settings, email, tokens } = await userService.getSettings();
-
-    if (email) {
-      commit(SET_EMAIL, email);
-    }
-
-    if (settings) {
-      commit(SET_SETTINGS, settings);
-    }
-
-    if (tokens) {
-      commit(`tokens/${SAVE_TOKENS}`, tokens, {
-        root: true,
-      });
-      // Saved token contract addresses on all networks
-      const tokenAddrs = []
-        .concat(...Object.values(tokens))
-        .map(token => token.address);
-      commit(`tokens/${SAVE_TRACKED_TOKENS}`, tokenAddrs, { root: true });
-    }
-  } catch (e) {
-    await dispatch('errors/emitError', e, { root: true });
-  }
+  return wallet.validatePassword(password);
 };
 
 const setUserHdKey = async ({ commit, dispatch }) => {
@@ -295,11 +216,7 @@ const setUserWallets = async ({ commit, dispatch }) => {
 
 const init = async ({ dispatch }) => {
   try {
-    await Promise.all([
-      dispatch('setUserSettings'),
-      dispatch('setUserHdKey'),
-      dispatch('setUserWallets'),
-    ]);
+    await Promise.all([dispatch('setUserHdKey'), dispatch('setUserWallets')]);
   } catch (e) {
     await dispatch('errors/emitError', e, { root: true });
   }
@@ -320,15 +237,5 @@ export default {
   addMultiHdWallet,
   updateBalance,
   validatePassword,
-
-  updateSettings,
-  login,
-  logout,
-  loginViaOTP,
-  getOtpSettings,
-  setOtpSettings,
-  setUserSettings,
-  deleteOtpSettings,
-
   init,
 };
