+++ src/store/accounts/index.js
@@ -3,8 +3,6 @@ import getters from './getters';
 import mutations from './mutations';
 
 const state = {
-  // TODO move to user store?
-  email: null,
   // The encrypted xprv key for the user's hd wallet as a V3 keystore
   // object. All accounts are direct child accounts.
   hdKey: null,
@@ -14,15 +12,6 @@ const state = {
   wallet: null,
   address: null,
   balance: null,
-  // prettier-ignore
-  availableCurrencies: ['USD', 'AUD', 'BRL', 'CAD', 'CHF', 'CLP', 'CNY', 'CZK', 'DKK', 'EUR', 'GBP', 'HKD', 'HUF', 'IDR', 'ILS', 'INR', 'JPY', 'KRW', 'MXN', 'MYR', 'NOK', 'NZD', 'PHP', 'PKR', 'PLN', 'RUB', 'SEK', 'SGD', 'THB', 'TRY', 'TWD', 'ZAR'],
-  settings: {
-    fiatCurrency: 'USD',
-  },
-  otpSettings: {
-    secret: null,
-    status: null,
-  },
 };
 
 export default {
