+++ src/store/accounts/mutations-types.js
@@ -4,9 +4,6 @@ export const ADD_WALLET = 'ADD_WALLET';
 export const SET_WALLET = 'SET_WALLET';
 export const SET_HD_KEY = 'SET_HD_KEY';
 export const SET_BALANCE = 'SET_BALANCE';
-export const SET_SETTINGS = 'SET_SETTINGS';
-export const SET_OTP_SETTINGS = 'SET_OTP_SETTINGS';
-export const SET_EMAIL = 'SET_EMAIL';
 
 export default {
   ADD_ADDRESS,
@@ -15,7 +12,4 @@ export default {
   SET_WALLET,
   SET_HD_KEY,
   SET_BALANCE,
-  SET_SETTINGS,
-  SET_OTP_SETTINGS,
-  SET_EMAIL,
 };
