+++ src/store/accounts/mutations.js
@@ -6,9 +6,6 @@ import {
   SET_WALLET,
   SET_HD_KEY,
   SET_BALANCE,
-  SET_SETTINGS,
-  SET_OTP_SETTINGS,
-  SET_EMAIL,
 } from './mutations-types';
 
 const setAddress = (state, newAddress) => {
@@ -52,18 +49,6 @@ const setBalance = (state, balance) => {
   state.balance = balance;
 };
 
-const setSettings = (state, settings) => {
-  state.settings = JSON.parse(JSON.stringify(settings));
-};
-
-const setOtpSettings = (state, otpSettings) => {
-  state.otpSettings = otpSettings;
-};
-
-const setEmail = (state, email) => {
-  state.email = email;
-};
-
 export default {
   [ADD_ADDRESS]: addAddress,
   [SET_ADDRESS]: setAddress,
@@ -71,7 +56,4 @@ export default {
   [SET_WALLET]: setWallet,
   [SET_HD_KEY]: setHdKey,
   [SET_BALANCE]: setBalance,
-  [SET_SETTINGS]: setSettings,
-  [SET_OTP_SETTINGS]: setOtpSettings,
-  [SET_EMAIL]: setEmail,
 };
