+++ src/store/price/getters.js
@@ -1,9 +1,7 @@
 const activeCurrencyName = (state, getters, { web3 }) => {
   return web3.activeCurrency.name;
 };
-const fiatCurrency = (state, getters, { accounts }) => {
-  return accounts.settings.fiatCurrency;
-};
+const fiatCurrency = (state, getters, { user }) => user.settings.fiatCurrency;
 
 export default {
   activeCurrencyName,
