+++ src/store/user/getters.js
@@ -1,7 +1,10 @@
+const isLoggedIn = state => !!state.email;
+
 const isLoggedOut = state =>
   !state.authorizationStatus &&
   state.authorizationStatus !== state.prevAuthorizationStatus;
 
 export default {
+  isLoggedIn,
   isLoggedOut,
 };
