+++ src/store/user/index.js
@@ -1,10 +1,20 @@
 import actions from './actions';
 import getters from './getters';
 import mutations from './mutations';
+import { AVAILABLE_FIAT_CURRENCIES } from '@/constants';
 
 const state = {
   prevAuthorizationStatus: null,
   authorizationStatus: null,
+  email: null,
+  settings: {
+    fiatCurrency: 'USD',
+  },
+  otpSettings: {
+    secret: null,
+    status: null,
+  },
+  availableCurrencies: AVAILABLE_FIAT_CURRENCIES,
 };
 
 export default {
