+++ src/store/user/mutations-types.js
@@ -1,1 +1,11 @@
 export const SET_AUTHORIZATION_STATUS = 'SET_AUTHORIZATION_STATUS';
+export const SET_EMAIL = 'SET_EMAIL';
+export const SET_SETTINGS = 'SET_SETTINGS';
+export const SET_OTP_SETTINGS = 'SET_OTP_SETTINGS';
+
+export default {
+  SET_AUTHORIZATION_STATUS,
+  SET_EMAIL,
+  SET_SETTINGS,
+  SET_OTP_SETTINGS,
+};
