+++ test/unit/specs/components/TokenList.spec.js
@@ -24,7 +24,7 @@ describe('TokenList', () => {
 
     store = new Vuex.Store({
       modules: {
-        accounts: {
+        user: {
           namespaced: true,
           state: {
             settings: {
