+++ test/unit/specs/components/modal/LoginModal.spec.js
@@ -85,7 +85,7 @@ describe('LoginModal', () => {
 
       const store = new Vuex.Store({
         modules: {
-          accounts: {
+          user: {
             namespaced: true,
             actions,
           },
