+++ test/unit/specs/components/pages/Tokens.spec.js
@@ -40,7 +40,7 @@ describe('TokensPage', () => {
 
     store = new Vuex.Store({
       modules: {
-        accounts: {
+        user: {
           namespaced: true,
           state: {
             settings: {
