+++ test/unit/specs/store/accounts/getters.spec.js
@@ -34,20 +34,6 @@ describe('Accounts getters', () => {
     });
   });
 
-  describe('isLoggedIn', () => {
-    it('should return false when not login', () => {
-      const state = { email: null };
-
-      expect(getters.isLoggedIn(state)).toBeFalsy();
-    });
-
-    it('should return true when login', () => {
-      const state = { email: 'email@email.com' };
-
-      expect(getters.isLoggedIn(state)).toBeTruthy();
-    });
-  });
-
   describe('balance', () => {
     it('should return null when the balance is null', () => {
       const state = { balance: null };
